Refactored COde to seperate Interface funtions into there own class
